+++ R/do_ExpressionHeatmap.R
@@ -147,7 +147,7 @@ do_ExpressionHeatmap <- function(sample,
                    crayon_key("list"),
                    crayon_body(" to the parameter "),
                    crayon_key("features"),
-                   crayon_body("Transforming into a character vector.")))
+                   crayon_body("Transforming into a character vector.")), call. = FALSE)
     features <- unname(unlist(features))
   }
 
@@ -305,10 +305,12 @@ do_ExpressionHeatmap <- function(sample,
     counter <- counter + 1
     data <- matrix.list[[group]][["data"]]
 
-    p <- ggplot2::ggplot(data,
-                         mapping = ggplot2::aes(x = if (isFALSE(flip)){.data$gene} else {.data$group.by},
+    p <- data %>% 
+         # nocov start
+         ggplot2::ggplot(mapping = ggplot2::aes(x = if (isFALSE(flip)){.data$gene} else {.data$group.by},
                                                 y = if (isFALSE(flip)){.data$group.by} else {.data$gene},
                                                 fill = .data$mean)) +
+         # nocov end
          ggplot2::geom_tile(color = grid.color, linewidth = 0.5) +
          ggplot2::scale_y_discrete(expand = c(0, 0)) +
          ggplot2::scale_x_discrete(expand = c(0, 0),